// src/.storyblok/types/339170/storyblok-components.d.ts
export interface MenuItem {
  Title?: string;
  link?: Exclude<StoryblokMultilink, { linktype?: "email" } | { linktype?: "asset" }>;
  SubItems?: SubMenu[];
  component: "menu_item";
  _uid: string;
  [k: string]: unknown;
}
// ...
export interface SubMenu {
  title?: string;
  link?: Exclude<StoryblokMultilink, { linktype?: "email" } | { linktype?: "asset" }>;
  component: "sub_menu";
  _uid: string;
  [k: string]: unknown;
}